use defaultTargetPlatform to check the platform instead of the platform package
