+++ flutter_local_notifications/test/android_flutter_local_notifications_test.dart
@@ -5,9 +5,10 @@ import 'package:flutter/services.dart';
 import 'package:flutter_local_notifications/flutter_local_notifications.dart';
 import 'package:flutter_local_notifications/src/platform_specifics/android/enums.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:platform/platform.dart';
 import 'package:timezone/data/latest.dart' as tz;
 import 'package:timezone/timezone.dart' as tz;
+
+import 'utils/date_formatter.dart';
 
 void main() {
   // TODO(maikub): add tests for `periodicallyShow` after https://github.com/dart-lang/sdk/issues/28985 is resolved
@@ -20,8 +21,8 @@ void main() {
     final List<MethodCall> log = <MethodCall>[];
 
     setUp(() {
-      flutterLocalNotificationsPlugin = FlutterLocalNotificationsPlugin.private(
-          FakePlatform(operatingSystem: 'android'));
+      debugDefaultTargetPlatformOverride = TargetPlatform.android;
+      flutterLocalNotificationsPlugin = FlutterLocalNotificationsPlugin();
       // ignore: always_specify_types
       channel.setMockMethodCallHandler((methodCall) async {
         log.add(methodCall);
@@ -1641,7 +1642,7 @@ void main() {
               'body': 'notification body',
               'payload': '',
               'timeZoneName': 'Australia/Sydney',
-              'scheduledDateTime': _convertDateToISO8601String(scheduledDate),
+              'scheduledDateTime': convertDateToISO8601String(scheduledDate),
               'platformSpecifics': <String, Object?>{
                 'allowWhileIdle': true,
                 'icon': null,
@@ -1731,7 +1732,7 @@ void main() {
               'body': 'notification body',
               'payload': '',
               'timeZoneName': 'Australia/Sydney',
-              'scheduledDateTime': _convertDateToISO8601String(scheduledDate),
+              'scheduledDateTime': convertDateToISO8601String(scheduledDate),
               'matchDateTimeComponents': DateTimeComponents.time.index,
               'platformSpecifics': <String, Object?>{
                 'allowWhileIdle': true,
@@ -1822,7 +1823,7 @@ void main() {
               'body': 'notification body',
               'payload': '',
               'timeZoneName': 'Australia/Sydney',
-              'scheduledDateTime': _convertDateToISO8601String(scheduledDate),
+              'scheduledDateTime': convertDateToISO8601String(scheduledDate),
               'matchDateTimeComponents':
                   DateTimeComponents.dayOfWeekAndTime.index,
               'platformSpecifics': <String, Object?>{
@@ -2088,774 +2089,4 @@ void main() {
           ));
     });
   });
-
-  group('iOS', () {
-    const MethodChannel channel =
-        MethodChannel('dexterous.com/flutter/local_notifications');
-    final List<MethodCall> log = <MethodCall>[];
-
-    setUp(() {
-      flutterLocalNotificationsPlugin = FlutterLocalNotificationsPlugin.private(
-          FakePlatform(operatingSystem: 'ios'));
-      // ignore: always_specify_types
-      channel.setMockMethodCallHandler((methodCall) async {
-        log.add(methodCall);
-        if (methodCall.method == 'pendingNotificationRequests') {
-          return <Map<String, Object>?>[];
-        } else if (methodCall.method == 'getNotificationAppLaunchDetails') {
-          return null;
-        }
-      });
-    });
-
-    tearDown(() {
-      log.clear();
-    });
-
-    test('initialize with default parameter values', () async {
-      const IOSInitializationSettings iosInitializationSettings =
-          IOSInitializationSettings();
-      const InitializationSettings initializationSettings =
-          InitializationSettings(iOS: iosInitializationSettings);
-      await flutterLocalNotificationsPlugin.initialize(initializationSettings);
-      expect(log, <Matcher>[
-        isMethodCall('initialize', arguments: <String, Object>{
-          'requestAlertPermission': true,
-          'requestSoundPermission': true,
-          'requestBadgePermission': true,
-          'defaultPresentAlert': true,
-          'defaultPresentSound': true,
-          'defaultPresentBadge': true,
-        })
-      ]);
-    });
-
-    test('initialize with all settings off', () async {
-      const IOSInitializationSettings iosInitializationSettings =
-          IOSInitializationSettings(
-              requestAlertPermission: false,
-              requestBadgePermission: false,
-              requestSoundPermission: false,
-              defaultPresentAlert: false,
-              defaultPresentBadge: false,
-              defaultPresentSound: false);
-      const InitializationSettings initializationSettings =
-          InitializationSettings(iOS: iosInitializationSettings);
-      await flutterLocalNotificationsPlugin.initialize(initializationSettings);
-      expect(log, <Matcher>[
-        isMethodCall('initialize', arguments: <String, Object>{
-          'requestAlertPermission': false,
-          'requestSoundPermission': false,
-          'requestBadgePermission': false,
-          'defaultPresentAlert': false,
-          'defaultPresentSound': false,
-          'defaultPresentBadge': false,
-        })
-      ]);
-    });
-
-    test('show without iOS-specific details', () async {
-      const IOSInitializationSettings iosInitializationSettings =
-          IOSInitializationSettings();
-      const InitializationSettings initializationSettings =
-          InitializationSettings(iOS: iosInitializationSettings);
-      await flutterLocalNotificationsPlugin.initialize(initializationSettings);
-      await flutterLocalNotificationsPlugin.show(
-          1, 'notification title', 'notification body', null);
-      expect(
-          log.last,
-          isMethodCall('show', arguments: <String, Object?>{
-            'id': 1,
-            'title': 'notification title',
-            'body': 'notification body',
-            'payload': '',
-            'platformSpecifics': null,
-          }));
-    });
-
-    test('show with iOS-specific details', () async {
-      const IOSInitializationSettings iosInitializationSettings =
-          IOSInitializationSettings();
-      const InitializationSettings initializationSettings =
-          InitializationSettings(iOS: iosInitializationSettings);
-      await flutterLocalNotificationsPlugin.initialize(initializationSettings);
-      const NotificationDetails notificationDetails = NotificationDetails(
-          iOS: IOSNotificationDetails(
-              presentAlert: true,
-              presentBadge: true,
-              presentSound: true,
-              subtitle: 'a subtitle',
-              sound: 'sound.mp3',
-              badgeNumber: 1,
-              attachments: <IOSNotificationAttachment>[
-            IOSNotificationAttachment('video.mp4',
-                identifier: '2b3f705f-a680-4c9f-8075-a46a70e28373'),
-          ]));
-
-      await flutterLocalNotificationsPlugin.show(
-          1, 'notification title', 'notification body', notificationDetails);
-
-      expect(
-          log.last,
-          isMethodCall('show', arguments: <String, Object>{
-            'id': 1,
-            'title': 'notification title',
-            'body': 'notification body',
-            'payload': '',
-            'platformSpecifics': <String, Object?>{
-              'presentAlert': true,
-              'presentBadge': true,
-              'presentSound': true,
-              'subtitle': 'a subtitle',
-              'sound': 'sound.mp3',
-              'badgeNumber': 1,
-              'threadIdentifier': null,
-              'attachments': <Map<String, Object>>[
-                <String, Object>{
-                  'filePath': 'video.mp4',
-                  'identifier': '2b3f705f-a680-4c9f-8075-a46a70e28373',
-                }
-              ],
-            },
-          }));
-    });
-
-    group('zonedSchedule', () {
-      test('no repeat frequency', () async {
-        const IOSInitializationSettings iosInitializationSettings =
-            IOSInitializationSettings();
-        const InitializationSettings initializationSettings =
-            InitializationSettings(iOS: iosInitializationSettings);
-        await flutterLocalNotificationsPlugin
-            .initialize(initializationSettings);
-        tz.initializeTimeZones();
-        tz.setLocalLocation(tz.getLocation('Australia/Sydney'));
-        final tz.TZDateTime scheduledDate =
-            tz.TZDateTime.now(tz.local).add(const Duration(seconds: 5));
-        const NotificationDetails notificationDetails = NotificationDetails(
-            iOS: IOSNotificationDetails(
-                presentAlert: true,
-                presentBadge: true,
-                presentSound: true,
-                sound: 'sound.mp3',
-                badgeNumber: 1,
-                attachments: <IOSNotificationAttachment>[
-              IOSNotificationAttachment('video.mp4',
-                  identifier: '2b3f705f-a680-4c9f-8075-a46a70e28373')
-            ]));
-
-        await flutterLocalNotificationsPlugin.zonedSchedule(
-            1,
-            'notification title',
-            'notification body',
-            scheduledDate,
-            notificationDetails,
-            androidAllowWhileIdle: true,
-            uiLocalNotificationDateInterpretation:
-                UILocalNotificationDateInterpretation.absoluteTime);
-
-        expect(
-            log.last,
-            isMethodCall('zonedSchedule', arguments: <String, Object>{
-              'id': 1,
-              'title': 'notification title',
-              'body': 'notification body',
-              'payload': '',
-              'uiLocalNotificationDateInterpretation':
-                  UILocalNotificationDateInterpretation.absoluteTime.index,
-              'scheduledDateTime': _convertDateToISO8601String(scheduledDate),
-              'timeZoneName': 'Australia/Sydney',
-              'platformSpecifics': <String, Object?>{
-                'presentAlert': true,
-                'presentBadge': true,
-                'presentSound': true,
-                'subtitle': null,
-                'sound': 'sound.mp3',
-                'badgeNumber': 1,
-                'threadIdentifier': null,
-                'attachments': <Map<String, Object>>[
-                  <String, Object>{
-                    'filePath': 'video.mp4',
-                    'identifier': '2b3f705f-a680-4c9f-8075-a46a70e28373',
-                  }
-                ],
-              },
-            }));
-      });
-
-      test('match time components', () async {
-        const IOSInitializationSettings iosInitializationSettings =
-            IOSInitializationSettings();
-        const InitializationSettings initializationSettings =
-            InitializationSettings(iOS: iosInitializationSettings);
-        await flutterLocalNotificationsPlugin
-            .initialize(initializationSettings);
-        tz.initializeTimeZones();
-        tz.setLocalLocation(tz.getLocation('Australia/Sydney'));
-        final tz.TZDateTime scheduledDate =
-            tz.TZDateTime.now(tz.local).add(const Duration(seconds: 5));
-        const NotificationDetails notificationDetails = NotificationDetails(
-            iOS: IOSNotificationDetails(
-                presentAlert: true,
-                presentBadge: true,
-                presentSound: true,
-                sound: 'sound.mp3',
-                badgeNumber: 1,
-                attachments: <IOSNotificationAttachment>[
-              IOSNotificationAttachment('video.mp4',
-                  identifier: '2b3f705f-a680-4c9f-8075-a46a70e28373')
-            ]));
-
-        await flutterLocalNotificationsPlugin.zonedSchedule(
-            1,
-            'notification title',
-            'notification body',
-            scheduledDate,
-            notificationDetails,
-            androidAllowWhileIdle: true,
-            uiLocalNotificationDateInterpretation:
-                UILocalNotificationDateInterpretation.absoluteTime,
-            matchDateTimeComponents: DateTimeComponents.time);
-
-        expect(
-            log.last,
-            isMethodCall('zonedSchedule', arguments: <String, Object>{
-              'id': 1,
-              'title': 'notification title',
-              'body': 'notification body',
-              'payload': '',
-              'uiLocalNotificationDateInterpretation':
-                  UILocalNotificationDateInterpretation.absoluteTime.index,
-              'scheduledDateTime': _convertDateToISO8601String(scheduledDate),
-              'timeZoneName': 'Australia/Sydney',
-              'matchDateTimeComponents': DateTimeComponents.time.index,
-              'platformSpecifics': <String, Object?>{
-                'presentAlert': true,
-                'presentBadge': true,
-                'presentSound': true,
-                'subtitle': null,
-                'sound': 'sound.mp3',
-                'badgeNumber': 1,
-                'threadIdentifier': null,
-                'attachments': <Map<String, Object>>[
-                  <String, Object>{
-                    'filePath': 'video.mp4',
-                    'identifier': '2b3f705f-a680-4c9f-8075-a46a70e28373',
-                  }
-                ],
-              },
-            }));
-      });
-
-      test('match day of week and time components', () async {
-        const IOSInitializationSettings iosInitializationSettings =
-            IOSInitializationSettings();
-        const InitializationSettings initializationSettings =
-            InitializationSettings(iOS: iosInitializationSettings);
-        await flutterLocalNotificationsPlugin
-            .initialize(initializationSettings);
-        tz.initializeTimeZones();
-        tz.setLocalLocation(tz.getLocation('Australia/Sydney'));
-        final tz.TZDateTime scheduledDate =
-            tz.TZDateTime.now(tz.local).add(const Duration(seconds: 5));
-        const NotificationDetails notificationDetails = NotificationDetails(
-            iOS: IOSNotificationDetails(
-                presentAlert: true,
-                presentBadge: true,
-                presentSound: true,
-                sound: 'sound.mp3',
-                badgeNumber: 1,
-                attachments: <IOSNotificationAttachment>[
-              IOSNotificationAttachment('video.mp4',
-                  identifier: '2b3f705f-a680-4c9f-8075-a46a70e28373')
-            ]));
-
-        await flutterLocalNotificationsPlugin.zonedSchedule(
-            1,
-            'notification title',
-            'notification body',
-            scheduledDate,
-            notificationDetails,
-            androidAllowWhileIdle: true,
-            uiLocalNotificationDateInterpretation:
-                UILocalNotificationDateInterpretation.absoluteTime,
-            matchDateTimeComponents: DateTimeComponents.dayOfWeekAndTime);
-
-        expect(
-            log.last,
-            isMethodCall('zonedSchedule', arguments: <String, Object>{
-              'id': 1,
-              'title': 'notification title',
-              'body': 'notification body',
-              'payload': '',
-              'uiLocalNotificationDateInterpretation':
-                  UILocalNotificationDateInterpretation.absoluteTime.index,
-              'scheduledDateTime': _convertDateToISO8601String(scheduledDate),
-              'timeZoneName': 'Australia/Sydney',
-              'matchDateTimeComponents':
-                  DateTimeComponents.dayOfWeekAndTime.index,
-              'platformSpecifics': <String, Object?>{
-                'presentAlert': true,
-                'presentBadge': true,
-                'presentSound': true,
-                'subtitle': null,
-                'sound': 'sound.mp3',
-                'badgeNumber': 1,
-                'threadIdentifier': null,
-                'attachments': <Map<String, Object>>[
-                  <String, Object>{
-                    'filePath': 'video.mp4',
-                    'identifier': '2b3f705f-a680-4c9f-8075-a46a70e28373',
-                  }
-                ],
-              },
-            }));
-      });
-    });
-    test('requestPermissions with default settings', () async {
-      await flutterLocalNotificationsPlugin
-          .resolvePlatformSpecificImplementation<
-              IOSFlutterLocalNotificationsPlugin>()!
-          .requestPermissions();
-      expect(log, <Matcher>[
-        isMethodCall('requestPermissions', arguments: <String, Object?>{
-          'sound': false,
-          'badge': false,
-          'alert': false,
-        })
-      ]);
-    });
-    test('requestPermissions with all settings requested', () async {
-      await flutterLocalNotificationsPlugin
-          .resolvePlatformSpecificImplementation<
-              IOSFlutterLocalNotificationsPlugin>()!
-          .requestPermissions(sound: true, badge: true, alert: true);
-      expect(log, <Matcher>[
-        isMethodCall('requestPermissions', arguments: <String, Object>{
-          'sound': true,
-          'badge': true,
-          'alert': true,
-        })
-      ]);
-    });
-    test('cancel', () async {
-      await flutterLocalNotificationsPlugin.cancel(1);
-      expect(log, <Matcher>[isMethodCall('cancel', arguments: 1)]);
-    });
-
-    test('cancelAll', () async {
-      await flutterLocalNotificationsPlugin.cancelAll();
-      expect(log, <Matcher>[isMethodCall('cancelAll', arguments: null)]);
-    });
-
-    test('pendingNotificationRequests', () async {
-      await flutterLocalNotificationsPlugin.pendingNotificationRequests();
-      expect(log, <Matcher>[
-        isMethodCall('pendingNotificationRequests', arguments: null)
-      ]);
-    });
-
-    test('getNotificationAppLaunchDetails', () async {
-      await flutterLocalNotificationsPlugin.getNotificationAppLaunchDetails();
-      expect(log, <Matcher>[
-        isMethodCall('getNotificationAppLaunchDetails', arguments: null)
-      ]);
-    });
-  });
-
-  group('macOS', () {
-    const MethodChannel channel =
-        MethodChannel('dexterous.com/flutter/local_notifications');
-    final List<MethodCall> log = <MethodCall>[];
-
-    setUp(() {
-      flutterLocalNotificationsPlugin = FlutterLocalNotificationsPlugin.private(
-          FakePlatform(operatingSystem: 'macos'));
-      // ignore: always_specify_types
-      channel.setMockMethodCallHandler((methodCall) async {
-        log.add(methodCall);
-        if (methodCall.method == 'pendingNotificationRequests') {
-          return <Map<String, Object>?>[];
-        } else if (methodCall.method == 'getNotificationAppLaunchDetails') {
-          return null;
-        }
-      });
-    });
-
-    tearDown(() {
-      log.clear();
-    });
-
-    test('initialize with default parameter values', () async {
-      const MacOSInitializationSettings macOSInitializationSettings =
-          MacOSInitializationSettings();
-      const InitializationSettings initializationSettings =
-          InitializationSettings(macOS: macOSInitializationSettings);
-      await flutterLocalNotificationsPlugin.initialize(initializationSettings);
-      expect(log, <Matcher>[
-        isMethodCall('initialize', arguments: <String, Object>{
-          'requestAlertPermission': true,
-          'requestSoundPermission': true,
-          'requestBadgePermission': true,
-          'defaultPresentAlert': true,
-          'defaultPresentSound': true,
-          'defaultPresentBadge': true,
-        })
-      ]);
-    });
-
-    test('initialize with all settings off', () async {
-      const MacOSInitializationSettings macOSInitializationSettings =
-          MacOSInitializationSettings(
-              requestAlertPermission: false,
-              requestBadgePermission: false,
-              requestSoundPermission: false,
-              defaultPresentAlert: false,
-              defaultPresentBadge: false,
-              defaultPresentSound: false);
-      const InitializationSettings initializationSettings =
-          InitializationSettings(macOS: macOSInitializationSettings);
-      await flutterLocalNotificationsPlugin.initialize(initializationSettings);
-      expect(log, <Matcher>[
-        isMethodCall('initialize', arguments: <String, Object>{
-          'requestAlertPermission': false,
-          'requestSoundPermission': false,
-          'requestBadgePermission': false,
-          'defaultPresentAlert': false,
-          'defaultPresentSound': false,
-          'defaultPresentBadge': false,
-        })
-      ]);
-    });
-
-    test('show without macOS-specific details', () async {
-      const MacOSInitializationSettings macOSInitializationSettings =
-          MacOSInitializationSettings();
-      const InitializationSettings initializationSettings =
-          InitializationSettings(macOS: macOSInitializationSettings);
-      await flutterLocalNotificationsPlugin.initialize(initializationSettings);
-      await flutterLocalNotificationsPlugin.show(
-          1, 'notification title', 'notification body', null);
-      expect(
-          log.last,
-          isMethodCall('show', arguments: <String, Object?>{
-            'id': 1,
-            'title': 'notification title',
-            'body': 'notification body',
-            'payload': '',
-            'platformSpecifics': null,
-          }));
-    });
-
-    test('show with macOS-specific details', () async {
-      const MacOSInitializationSettings macOSInitializationSettings =
-          MacOSInitializationSettings();
-      const InitializationSettings initializationSettings =
-          InitializationSettings(macOS: macOSInitializationSettings);
-      await flutterLocalNotificationsPlugin.initialize(initializationSettings);
-      const NotificationDetails notificationDetails = NotificationDetails(
-          macOS: MacOSNotificationDetails(
-              subtitle: 'a subtitle',
-              presentAlert: true,
-              presentBadge: true,
-              presentSound: true,
-              sound: 'sound.mp3',
-              badgeNumber: 1,
-              threadIdentifier: 'thread',
-              attachments: <MacOSNotificationAttachment>[
-            MacOSNotificationAttachment('video.mp4',
-                identifier: '2b3f705f-a680-4c9f-8075-a46a70e28373'),
-          ]));
-
-      await flutterLocalNotificationsPlugin.show(
-          1, 'notification title', 'notification body', notificationDetails);
-
-      expect(
-          log.last,
-          isMethodCall('show', arguments: <String, Object>{
-            'id': 1,
-            'title': 'notification title',
-            'body': 'notification body',
-            'payload': '',
-            'platformSpecifics': <String, Object>{
-              'subtitle': 'a subtitle',
-              'presentAlert': true,
-              'presentBadge': true,
-              'presentSound': true,
-              'sound': 'sound.mp3',
-              'badgeNumber': 1,
-              'threadIdentifier': 'thread',
-              'attachments': <Map<String, Object>>[
-                <String, Object>{
-                  'filePath': 'video.mp4',
-                  'identifier': '2b3f705f-a680-4c9f-8075-a46a70e28373',
-                }
-              ],
-            },
-          }));
-    });
-
-    group('zonedSchedule', () {
-      test('no repeat frequency', () async {
-        const MacOSInitializationSettings macOSInitializationSettings =
-            MacOSInitializationSettings();
-        const InitializationSettings initializationSettings =
-            InitializationSettings(macOS: macOSInitializationSettings);
-        await flutterLocalNotificationsPlugin
-            .initialize(initializationSettings);
-        tz.initializeTimeZones();
-        tz.setLocalLocation(tz.getLocation('Australia/Sydney'));
-        final tz.TZDateTime scheduledDate =
-            tz.TZDateTime.now(tz.local).add(const Duration(seconds: 5));
-        const NotificationDetails notificationDetails = NotificationDetails(
-            macOS: MacOSNotificationDetails(
-                presentAlert: true,
-                presentBadge: true,
-                presentSound: true,
-                sound: 'sound.mp3',
-                badgeNumber: 1,
-                attachments: <MacOSNotificationAttachment>[
-              MacOSNotificationAttachment('video.mp4',
-                  identifier: '2b3f705f-a680-4c9f-8075-a46a70e28373')
-            ]));
-
-        await flutterLocalNotificationsPlugin.zonedSchedule(
-            1,
-            'notification title',
-            'notification body',
-            scheduledDate,
-            notificationDetails,
-            androidAllowWhileIdle: true,
-            uiLocalNotificationDateInterpretation:
-                UILocalNotificationDateInterpretation.absoluteTime);
-
-        expect(
-            log.last,
-            isMethodCall('zonedSchedule', arguments: <String, Object>{
-              'id': 1,
-              'title': 'notification title',
-              'body': 'notification body',
-              'payload': '',
-              'scheduledDateTime': _convertDateToISO8601String(scheduledDate),
-              'timeZoneName': 'Australia/Sydney',
-              'platformSpecifics': <String, Object?>{
-                'subtitle': null,
-                'presentAlert': true,
-                'presentBadge': true,
-                'presentSound': true,
-                'sound': 'sound.mp3',
-                'badgeNumber': 1,
-                'threadIdentifier': null,
-                'attachments': <Map<String, Object>>[
-                  <String, Object>{
-                    'filePath': 'video.mp4',
-                    'identifier': '2b3f705f-a680-4c9f-8075-a46a70e28373',
-                  }
-                ],
-              },
-            }));
-      });
-
-      test('match time components', () async {
-        const MacOSInitializationSettings macOSInitializationSettings =
-            MacOSInitializationSettings();
-        const InitializationSettings initializationSettings =
-            InitializationSettings(macOS: macOSInitializationSettings);
-        await flutterLocalNotificationsPlugin
-            .initialize(initializationSettings);
-        tz.initializeTimeZones();
-        tz.setLocalLocation(tz.getLocation('Australia/Sydney'));
-        final tz.TZDateTime scheduledDate =
-            tz.TZDateTime.now(tz.local).add(const Duration(seconds: 5));
-        const NotificationDetails notificationDetails = NotificationDetails(
-            macOS: MacOSNotificationDetails(
-                presentAlert: true,
-                presentBadge: true,
-                presentSound: true,
-                sound: 'sound.mp3',
-                badgeNumber: 1,
-                attachments: <MacOSNotificationAttachment>[
-              MacOSNotificationAttachment('video.mp4',
-                  identifier: '2b3f705f-a680-4c9f-8075-a46a70e28373')
-            ]));
-
-        await flutterLocalNotificationsPlugin.zonedSchedule(
-            1,
-            'notification title',
-            'notification body',
-            scheduledDate,
-            notificationDetails,
-            androidAllowWhileIdle: true,
-            uiLocalNotificationDateInterpretation:
-                UILocalNotificationDateInterpretation.absoluteTime,
-            matchDateTimeComponents: DateTimeComponents.time);
-
-        expect(
-            log.last,
-            isMethodCall('zonedSchedule', arguments: <String, Object>{
-              'id': 1,
-              'title': 'notification title',
-              'body': 'notification body',
-              'payload': '',
-              'scheduledDateTime': _convertDateToISO8601String(scheduledDate),
-              'timeZoneName': 'Australia/Sydney',
-              'matchDateTimeComponents': DateTimeComponents.time.index,
-              'platformSpecifics': <String, Object?>{
-                'subtitle': null,
-                'presentAlert': true,
-                'presentBadge': true,
-                'presentSound': true,
-                'sound': 'sound.mp3',
-                'badgeNumber': 1,
-                'threadIdentifier': null,
-                'attachments': <Map<String, Object>>[
-                  <String, Object>{
-                    'filePath': 'video.mp4',
-                    'identifier': '2b3f705f-a680-4c9f-8075-a46a70e28373',
-                  }
-                ],
-              },
-            }));
-      });
-
-      test('weekly repeat frequency', () async {
-        const MacOSInitializationSettings macOSInitializationSettings =
-            MacOSInitializationSettings();
-        const InitializationSettings initializationSettings =
-            InitializationSettings(macOS: macOSInitializationSettings);
-        await flutterLocalNotificationsPlugin
-            .initialize(initializationSettings);
-        tz.initializeTimeZones();
-        tz.setLocalLocation(tz.getLocation('Australia/Sydney'));
-        final tz.TZDateTime scheduledDate =
-            tz.TZDateTime.now(tz.local).add(const Duration(seconds: 5));
-        const NotificationDetails notificationDetails = NotificationDetails(
-            macOS: MacOSNotificationDetails(
-                presentAlert: true,
-                presentBadge: true,
-                presentSound: true,
-                sound: 'sound.mp3',
-                badgeNumber: 1,
-                attachments: <MacOSNotificationAttachment>[
-              MacOSNotificationAttachment('video.mp4',
-                  identifier: '2b3f705f-a680-4c9f-8075-a46a70e28373')
-            ]));
-
-        await flutterLocalNotificationsPlugin.zonedSchedule(
-            1,
-            'notification title',
-            'notification body',
-            scheduledDate,
-            notificationDetails,
-            androidAllowWhileIdle: true,
-            uiLocalNotificationDateInterpretation:
-                UILocalNotificationDateInterpretation.absoluteTime,
-            matchDateTimeComponents: DateTimeComponents.dayOfWeekAndTime);
-
-        expect(
-            log.last,
-            isMethodCall('zonedSchedule', arguments: <String, Object>{
-              'id': 1,
-              'title': 'notification title',
-              'body': 'notification body',
-              'payload': '',
-              'scheduledDateTime': _convertDateToISO8601String(scheduledDate),
-              'timeZoneName': 'Australia/Sydney',
-              'matchDateTimeComponents':
-                  DateTimeComponents.dayOfWeekAndTime.index,
-              'platformSpecifics': <String, Object?>{
-                'subtitle': null,
-                'presentAlert': true,
-                'presentBadge': true,
-                'presentSound': true,
-                'sound': 'sound.mp3',
-                'badgeNumber': 1,
-                'threadIdentifier': null,
-                'attachments': <Map<String, Object>>[
-                  <String, Object>{
-                    'filePath': 'video.mp4',
-                    'identifier': '2b3f705f-a680-4c9f-8075-a46a70e28373',
-                  }
-                ],
-              },
-            }));
-      });
-    });
-
-    test('requestPermissions with default settings', () async {
-      await flutterLocalNotificationsPlugin
-          .resolvePlatformSpecificImplementation<
-              MacOSFlutterLocalNotificationsPlugin>()!
-          .requestPermissions();
-      expect(log, <Matcher>[
-        isMethodCall('requestPermissions', arguments: <String, Object?>{
-          'sound': null,
-          'badge': null,
-          'alert': null,
-        })
-      ]);
-    });
-    test('requestPermissions with all settings requested', () async {
-      await flutterLocalNotificationsPlugin
-          .resolvePlatformSpecificImplementation<
-              MacOSFlutterLocalNotificationsPlugin>()!
-          .requestPermissions(sound: true, badge: true, alert: true);
-      expect(log, <Matcher>[
-        isMethodCall('requestPermissions', arguments: <String, Object>{
-          'sound': true,
-          'badge': true,
-          'alert': true,
-        })
-      ]);
-    });
-    test('cancel', () async {
-      await flutterLocalNotificationsPlugin.cancel(1);
-      expect(log, <Matcher>[isMethodCall('cancel', arguments: 1)]);
-    });
-
-    test('cancelAll', () async {
-      await flutterLocalNotificationsPlugin.cancelAll();
-      expect(log, <Matcher>[isMethodCall('cancelAll', arguments: null)]);
-    });
-
-    test('pendingNotificationRequests', () async {
-      await flutterLocalNotificationsPlugin.pendingNotificationRequests();
-      expect(log, <Matcher>[
-        isMethodCall('pendingNotificationRequests', arguments: null)
-      ]);
-    });
-
-    test('getNotificationAppLaunchDetails', () async {
-      await flutterLocalNotificationsPlugin.getNotificationAppLaunchDetails();
-      expect(log, <Matcher>[
-        isMethodCall('getNotificationAppLaunchDetails', arguments: null)
-      ]);
-    });
-  });
-}
-
-String _convertDateToISO8601String(tz.TZDateTime dateTime) {
-  String _twoDigits(int n) {
-    if (n >= 10) {
-      return '$n';
-    }
-    return '0$n';
-  }
-
-  String _fourDigits(int n) {
-    final int absN = n.abs();
-    final String sign = n < 0 ? '-' : '';
-    if (absN >= 1000) {
-      return '$n';
-    }
-    if (absN >= 100) {
-      return '${sign}0$absN';
-    }
-    if (absN >= 10) {
-      return '${sign}00$absN';
-    }
-    return '${sign}000$absN';
-  }
-
-  return '${_fourDigits(dateTime.year)}-${_twoDigits(dateTime.month)}-${_twoDigits(dateTime.day)}T${_twoDigits(dateTime.hour)}:${_twoDigits(dateTime.minute)}:${_twoDigits(dateTime.second)}'; // ignore: lines_longer_than_80_chars
 }
