+++ flutter_local_notifications/test/utils/date_formatter.dart
@@ -1,0 +1,27 @@
+import 'package:timezone/timezone.dart' as tz;
+
+String convertDateToISO8601String(tz.TZDateTime dateTime) {
+  String _twoDigits(int n) {
+    if (n >= 10) {
+      return '$n';
+    }
+    return '0$n';
+  }
+
+  String _fourDigits(int n) {
+    final int absN = n.abs();
+    final String sign = n < 0 ? '-' : '';
+    if (absN >= 1000) {
+      return '$n';
+    }
+    if (absN >= 100) {
+      return '${sign}0$absN';
+    }
+    if (absN >= 10) {
+      return '${sign}00$absN';
+    }
+    return '${sign}000$absN';
+  }
+
+  return '${_fourDigits(dateTime.year)}-${_twoDigits(dateTime.month)}-${_twoDigits(dateTime.day)}T${_twoDigits(dateTime.hour)}:${_twoDigits(dateTime.minute)}:${_twoDigits(dateTime.second)}'; // ignore: lines_longer_than_80_chars
+}
